Recursive evaluation.
This makes the behavior more like Protovis, where the actions associated with a
given selection are processed recursively. This allows us to build a data stack,
so that the parent data is accessible to property functions. I haven't quite
figured out how I'm going to expose indexes yet, though. And still up in the air
is how the child data is extracted from parent data.