Fold key and index into data action.
The previous index action is now the default behavior when data is bound to the
selection. This is handy because it matches the default assignment of data and
means that `enter` and `exit` hooks are immediately available. The `key` method
now modifies the behavior of the current data binding, rather than creating a
new scope.